Altered the way parcelsetup works by adding the possibility to read in from existing files, or a custom setup routine.
When reading in existing files, they need not have been created by the same
number of processes that they are being read in from, allowing more
flexibility when running pmpic. Each process will only open the files it needs
in order to read all the parcels in its domain in.

When using a custom setup routine, the programmer puts the fortran source in
components/parcelsetup/src/, includes its module in components/src/parcelsetup.F90
and gives it a name which is put into the config file under "initialisation_routine".
As an example I have the "basic" routine included, which uniformly places
parcels in the computaional domain.